requirements.txt
uvicorn==0.18.2
RPi.bme280==0.2.4
smbus2==0.4.2
fastapi==0.79.0
pydantic==1.9.1
